Remove double price for the same token in clearing prices (bh2smith/cow-solver-rewards#21)
I found an ugly bug: If the clearing price is not unique in the clearing price table, it will duplicate many entries and will find too many internal trades. This caused our slippage results to be for quasimodo.

One example is the tx from the new test. It happened only for settlements with liquidity orders and hence it was hard to find.

Testplan:
Understand the unit test. Make sure the prices from the clearing price table are unique.