3. Redirect: COW Rewards (#21)
As part of the 3 step process this is the final PR introducing COW reward redirects. The script responsible for constructing the transfer file, intercepts any COW token transfers and replaces the `recipient` with the `reward_target` specified via `VouchRegistry`. You will notice that if there is nor specified redirect, the recipient is left as is (i.e. the solver itself).

To save gas, It would be possible to "consolidate" transfers of the same token to the same recipient by adding them up together. We have implemented transfer consolidation here.