Local Testing IV: Migrate VouchRegistry Tests (#47)
This is the biggest test migration because this is the first test query where we actually unit test with our own input records. This PR introduces a first working model for a framework that will allow Dune Query Unit testing the way it should be done (i.e. without relying on slices of Dune Data).