Refactor Transfer Part 1: Use Token Field (instead of token_address) (#71)
For the upcoming PRs related to MultiSend #66 & #67 we realized that we need to have a better approach for converting amount_wei to amount. This involves knowing the token decimals. This PR introduces a Token class (which holds both the token address and decimals).

The constructor of the token class has an optional decimals field that can be provided for unit testing, but when not available is fetched via get_token_decimals (i.e. making a third-party web3 call). We also avoid the 3rd party web3 call for the COW token (since it is the primary erc20 token used in this project) by hard-coding its decimals to be 18.

The Transfer Class has been changed to have a token field now instead of token address.
This wound up being quite a big change overall (i.e. many unit tests had to be refactored).