Refactor Transfer Part 2: Remove TokenType (#72)
Previously we had two coupled fields TokenType and Token with the following coupled logic:

Token = None <==> TokenType is NATIVE
Token != None <==> TokenType is ERC20

There was no real reason for these assertions. The only reason we had this here was because the CSVTransfer requires token type. So, we keep the token type as a property of the Transfer class used for filtering and also when converted to CSVTransfer.

Furthermore, there were special constructor methods Transfer.native and Transfer.erc20 that were only being used for unit tests. For the purpose of making the actual code shorter, these are removed.