[Automation] Default Start Date (#82)
Right now we manually pass in a start date parameter to this script. However, during automation it will be easier for the script to dynamically know the accounting period to run for. Thus, we make this argument default to the previous Tuesday. Since the job will run on Tuesdays, the deployed script will use rely on the default (previous Tuesday) as its Accounting Period Start.

Note: Additionally, that this change does not affect the current manual process as described in the README.