Preparing for Risk Adjusted Rewards (#89)
As a first step, we separate the eth_spent from the cow rewards dune query. Making independent fetching methods so that the get_cow_rewards can utilize a different source when the time comes. Currently this change does not affect the existing situation and can be merged.

Note that we have added V2 versions of the queries (commented out in the code, but ready for the switch if/when necessary).

In a follow up PR we will put together the expected query and connection to the orderbook DB.