[Cleanup] Separate Transfer File (#122)
The transfer file became quite large as the central script for generating solver rewards. This PR moves all the classes and independent functionality into appropriate models directory and relieves the fetch/transfer_file.py script of a lot of heavy weight.

Only inconvenience, is that we now had to pass the log_saver (as an instance of PrintStore) into some of the methods that are responsible for processing payouts.