Step 1.5: Partially Remove Legacy Dune (#132)
As an intermediary hybrid, we temporarily add both dune clients to the DuneFetcher and use the REAL dune client for almost everything. Exceptions being:

    period_slippage:
    vouch_registry

The reason for temporarily leaving this is because of the way these query has been broken up for different purposes - I will have to be a bit more careful when attempting to transition this and make sure not to miss anything.

Additionally, we introduce a Query "Tracker" File Which consists of an account of all queries contained in the project (both v1 and v2 versions where applicable). This PR does not yet change the queries being executed (i.e. they are all still the V1 versions of the query), but we will be able to transition to V2 with the flip of a switch, before Flipping this switch, however, I would like to completely remove the duneapi legacy client.

These more intricate cases will be saved for a follow up PR.