Add Existing V2 Queries (#133)
This PR merely adds V2 queries that have been migrated (in order to give a sense of completeness to the previous).

We do not yet make use of these or test them (however I have manually checked that they agree). I was getting overwhelmed with the amount of stray code I had sitting around and just wanted to get it in here.

The only missing queries for migration are period_slippage and period_totals. In the near future, we will introduce a form of "cross-check" between both versions of the queries.