[Easy] Add Makefile (#136)
Contains a bunch of convenience items. Specifically check and test-all which do all the things. Something which is also nice, but could be better is the db tests which checks if the db container image is running before the tests (it also builds and runs the image if not running).