Migrate Reward Targets (aka Vouch Registry) (#143)
This was actually the last query to be executed using the legacy-api (however we have not fully removed our dependency on it because of the Dashboard manager). In a follow up PR (immediately after this) I will remove legacy dune client from the DuneFetcher Object. Afterward, we proceed with testing all the v1<>v2 query results.

Uses dune-client for vouch registry, updates corresponding db-tests (after slight modification of the v1 query parameters). Adds a test that both v1 and v2 queries produce identical results.