*Properly Testing Slippage Query (#148)
Our previous test was not good enough and several issues were found (related to the minor changes made here). we are still not entirely accurate on the dex_swaps field (a crucial value for this calculation) because v2's dex.trades table is not fully populated with all the supported liquidity sources available. Furthermore, there was a logical constraint in one query that was not syntactically representative of its actual meaning.

We now test the slippage query at several "checkpoints" to ensure records are matching. Unfortunately, it is clear that our record filtration cannot be 100% accurate at this stage due to the lack of dex.trades precision. However, we do make up for it later, because the price feed in V2 is more complete.