[V1] Add surplus_fee to slippage query  (#156)
Separating #155 into two "independent" PRs so we can add tests n stuff.

The slippage query has been altered to account for surplus_fee. A new test is added, demonstrating that this surplus fee is captured by the V1 query and not the V2 query (since this is not yet possible).

Note that while testing this, I have discovered a bug in our base slippage query. Receiver = 0 There are only 113 total trades over all time with this set (see here) which had to be fixed before we could properly test this. That is because the query thought that there was an AMM_OUT transfer.