[Easy] Default start date change (#173)
We are now running this service as a "dry-run" on mondays and the current default of "Previous Tuesday" messes up the Monday run since this day hasn't happened yet. Instead we use default start date as today - 7 days which will always work.

Also removed `previous_tuesday` function (as it is no longer used) and its tests.