Disable Transfer Consolidation (#194)
We will spend more gas to do individual transfers for all Rewards, Gas Reimbursements and Positive slippage. This is supposed to make it easier to validate the solver payouts and in some sense it kinda does. The records are sorted in the same way they will appear in the dashboard.