[CIP-20] Unit Testing Reward Query (#200)
Based on #199

Writing code to execute reward extraction queries on both DBs. This introduces some local setup for unit testing the query. In order to avoid lopsided expectation that we can execute the same query on both we introduce an abstraction for having an array of database connections on which we can execute the same query an return the merged results.