Fetch: Winning Settlement from Orderbook API (#220)
A standalone component of the internalized transfer workflow, we fetch the WinningSettlement Data from the Orderbook API. This PR adds a model for the relevant data, a fetching method and tests for all possible cases. Remains in Draft State because we may want to use txReceipt instead of txHash as function input.