[Easy] Aggregate Transfers (#233)
Part of #177

A preliminary step before #231. In order to compute token imbalances we will need to aggregate collections of transfers by Token around a "focalAddress" (the settlement contract). This is just a utility accumulation method, a slight type change, add a relevant field to TransferEvent that was missing (namely token).