🔑 - Get Internalized Imbalance (#231)
This is a key component of the overall slippage accuracy project.

Implementation of getInternalImbalance which takes a transaction event and computes the internalized token imbalance. This is done by fetching the "full" and "reduced" callData from the winning solution (via Orderbook API), simulating this call data, parsing/aggregating the transfers and taking the difference of these with the actual transaction.