[Refactor] Split getInternalImbalance (#250)
As a result of yesterdays sync meeting, it appears we will want to "write Simulations to the database" previously the simulations only lived within the scope of the aggregation method, so we pull this logic out in order to have the simulation result object in our hands. This will enable us to write results to DB as we acquire them (i.e. before any post processing is done).

This will be followed up with a DB schema change and Insert Methods for the new struct introduced here (SettlementSimulationData).