[DB] Insert Pipeline Results as Single DB Transaction (#252)
After our pipeline has succeeded, we want to write to three different tables in a fail safe way (i.e. all or nothing). For this purpose we implement a method that commits all three insertions to the DB in such a fail safe way -- as a single database transaction.

Note that we had to remove the "insertOrIgnore" functionality on a few tables because we do not want such operations to "fail silently".