Simulate Both (#270)
It is well known that the WinningSettlement results returned from the
orderbook API have an unreliable `simulationBlock`. IN order to improve
our accuracy, we attempt to simulate the full and reduced settlement
call data on `simulationBlock + i` for `i = {0, 1, 2}`. This PR
introduces the necessary data structures, retry functionality and an
"e2e" test demonstrating the need for it.

---------

Co-authored-by: Gent Rexha <gent@cow.fi>