Preliminary & Primary Pipeline Functionality (#272)
This is best summed up by the following story (which should probably be
included somewhere in our documentation).

# The life cycle of our settlement event action

## Preliminary Pipeline

At the moment a Settlement Transaction is detected, the action is
triggered. The primary pipeline requires availability of the competition
data, in particular, that of the winning settlement. However, this data
is not yet available in the orderbook to be queried by transaction hash
because the block containing this transaction has not yet finalized. The
orderbook is configured to treat 65 block confirmations as finlization,
so our process must wait at least 65 blocks. To circumvent the
inconveniences of block finalization, the settlement transaction data is
"stashed" in the database as unprocessed, raw transaction receipts. A
secondary functionality of the pipeline then fetches all unprocessed
settlement transactions with `blockNumber > currentBlock - 65` (these
are assumed to be finalized) and executes the primary pipeline described
as follows:

## Primary Pipeline

Assuming availablity of the settlement transaction receipt & solver
competition data, this process will parse winning settlement for its
full and reduced call data. Both variants of the settlment are simulated
and disected for transfer ledgers. An aggregate difference of these
transfer ledgers is computed and written to the database along with the
simultion results.


This PR implements both of the above methods - utilizing all the tools
build in this project thus far.

---------

Co-authored-by: Gent Rexha <gent@cow.fi>