[Bug] Log Index Parsing (#275)
## Explanation of Bug

A previous deployment encountered a [primary-key
error](https://jsonblob.com/1106009775560474624) on insertion of
settlements with primary key `(blockNumber, LogIndex)` and this was
entirely due to the fact that we had been using the transaction-wise log
index as opposed to the block-wise log index.

### Example: 

1. Passing Execution
[c125a325-69c4-4acc-8b9d-dc0609eb0143](https://dashboard.tenderly.co/gp-v2/solver-slippage/action/4f657ddd-c054-4704-9e1e-50385ad2fc2c/execution/c125a325-69c4-4acc-8b9d-dc0609eb0143)
for tx

[0xe08fd9651626cc0827a83721e9b6ef99a8be752d2a88490218f75ba84082a887](https://etherscan.io/tx/0xe08fd9651626cc0827a83721e9b6ef99a8be752d2a88490218f75ba84082a887)

and

2. Failing Execution
[8d30189b-7282-4f63-990a-7d6849c49e62](https://dashboard.tenderly.co/gp-v2/solver-slippage/action/4f657ddd-c054-4704-9e1e-50385ad2fc2c/execution/8d30189b-7282-4f63-990a-7d6849c49e62)
for tx
[0xec894e53627f4b0a05be069a6c0c0cca08efe67f0b12deb85aeaa146f77eb049](https://etherscan.io/tx/0xec894e53627f4b0a05be069a6c0c0cca08efe67f0b12deb85aeaa146f77eb049)

where both settlements occurred in the same block and had exactly the
same settlement event index of 12 (transaction-wise) clashed here.
Meanwhile the block- wise log index of these events were 104 and 227
respectively.

Unfortunately, Tenderly does not provide the correct log index in their
[web-action payloads](http://jsonblob.com/1103717033526444032) nor in
their simulation results... so we are forced to fetch [transaction
receipts](http://jsonblob.com/1103716885064859648) from the EVM
(basically ignoring the entire payload data except for txHash)


## Solution

Here we introduce a field `index` on our model `EventLog` and populate
it with the values contained in `TransactionReceipt`. Having imposed
this in our work flow detected several places where the field was
required, but not yet available. In particular we are forces to "fudge"
the values on the Tenderly Simulation results because they are not
provided and they are only hypothetical transactions anyway. This does
not affect us in any way because we do not write Settlement Event data
for simulated transactions anyway -- furthermore, since we always
simulate at slot 0, these fudged values are actually the correct log
indices.


## Test Plan

All relevant unit tests have been updated. Along with a couple more for
the example provided above.


This is nearly the final PR for Phase 1 of #177