Insert Settlement & Mark Processed (#276)
Part of #177 

Doing a TODO to mark data as process when simulation is inserted. This
happens in both cases of the pipeline (when there is an is not a
simulation to go with it).

## Test Plan

Every component is tested individually. Could also add a test fot
`insertSettlementAndMarkProcessed` although it just runs both, tested
methods, together.