[Bug] Catch Missing Competition (#281)
Closes #280 

We had a short outage due to a forked block / failed transaction (cf
[tenderly](https://dashboard.tenderly.co/gp-v2/solver-slippage/action/4f657ddd-c054-4704-9e1e-50385ad2fc2c/execution/a14e3e0f-3e26-4290-8684-1720d74f1513)).

This
[transaction](https://etherscan.io/tx/0x1231f2c9b519adc5ae7db17c84418e140553e234b2868d6b1d7f66a692683e73)
was previously included in a block that wound up forked (tenderly picked
this up as successful beforehand) we stored the original successful
transaction in our DB, but then later it turns out to be a failed
transaction. This explains why the solver competition was not found.

Instead of throwing an Error when competition is not found, we can at
least go back to the EVM and check if the transaction details still
match what we have. I am not exactly sure how to uniquely identify a
"failed" transaction, but it suffices for our purposes that the
transaction does not have any logs.


## Test Plan

Added a test demonstrating that the pipeline does not throw when this
scenario occurs.