Add Enso Simulator (#284)
We recently started experimenting with [this transaction
simulator](https://github.com/EnsoFinance/transaction-simulator) that
requires only an Alchemy Archive Node and it works very well! Here we
implement the TransactionSimulator interface with an Enso Simulator.

This could be used to back fill our records rather than paying for over
100k tenderly simulations. This is exactly what I intend to do with it.
This will go well in combination with
https://github.com/cowprotocol/solver-rewards/pull/260

adds the first Alternative Simulator -- closes #245


### Test Plan
1. Run a local simulation API (as instructed in Enso) with default
localhost and port
2. remove `.skip` from the e2e test here and `npm run test` will pick up
the enso tests.