[DB] Add Primary Key to TxReceipts Table (#288)
We had some duplicated records in the tx receipt table every run of the
tenderly action was always trying to process this same transaction every
time.

This can occur if tenderly triggers an action for a transaction hash
twice or if we trigger the actions manually. Luckily we are still in
staging and the tx_receipts table is not guaranteed to contain all
records. We also aim to mark processed when we discover event record
existence.


The actual correct way to apply these kinds of changes are to create a
new version and run `ALTER TABLE... ` -- However, since we are not in
production yet, I don't see the point.

Locally (in my test DB), the table was fixed with:

```sql
-- deduplicate
delete from tx_receipts 
where hash in (select hash from tx_receipts group by hash having count(*) > 1);
-- add primary key
ALTER table tx_receipts add primary key (hash);
```