[Slippage] Account for Phantom Transfers (#259)
This is the Final piece of the new slippage accounting project. We
update the Slippage query to make use of the phantom transfers table.

In the query we see a URL to [this PoC
Query](https://dune.com/queries/2421375?StartTime_d83555=2023-01-01+00%3A00%3A00&EndTime_d83555=2023-01-02+00%3A00%3A00&CTE_NAME_e15077=incoming_and_outgoing_with_internalized_imbalances&sidebar=none)
with some manually set values (just to demonstrate that it works). We
need to populate the community sources with dune-sync before we can
actually land this.

BTW -- It looks like this will also improve our query performance (from
4 minutes down to 30 seconds).

Note that this PR only introduces a change to the query, but not
actually making use of it yet in the payout script.