Documenting Solver Slippage Accounting. (#302)
This documentation may not wind up living here (this is why I did not
use internal project references in the links). For the most recent
development regarding phantom transfers please see the dedicated section
for it.

This can, of course, be expanded in many directions, I just wanted to
get something out there today.