Ignore ETH wrapping transfers (#322)
Before only transfers from WETH to the contract were ignored. Those
correspond to unwraps/withdrawals.
With this change, also transfers from the protocol to WETH are ignored.
Those correspond to wraps/deposits.

The general idea is that all transfers between the protocol and WETH
should correspond to wrapping and unwrapping. Such operations do change the ETH balance and WETH balance without incurring any slippage or change in value of the buffers.