Fix query id for slippage accounting (#323)
The change to the `period_slippage.sql` query in #322 did not change the
dune query called from the code. This resulted in wrong slippage
computations from this repository. The Dune dashboards already showed
the correct slippage.
In this PR, the code in `queries.py` is changed to point to a new Dune
query created by cowprotocol. The code of that query is identical to
`period_slippage.sql`.