change all occurrences of BETWEEN to explicit inequalities
Instead of between, we always use a half-open interval which does not
include the left boundary but does include the right boundary. With this
approach, double counting between adjacent weeks is avoided.
For some tables, it does not matter if boundaries are included. In most
cases we stick to the above convention. There are exceptions for computing
mean prices where both bounds are included. There, double counting does
not matter.