Make block ranges consistent (take2) (#330)
This PR is based heavily on PR #329 but proposes a slightly different
convention. Specifically, it assumes that we compute a `start` and `end`
block for the accounting period, and both `start` and `end` block are
included in the accounting, as it seems more natural to the author.

E.g., this means that for consecutive weeks, we have

week 1: [a,b]
week 2: [b+1, c]
week 3: [c+1, d]

etc.

All queries (except for the vouch registry one) have been updated in the
form of forks. The plan would be to test this branch locally in the
upcoming payouts of Jan 9, and compare results (or alternatively, use
this branch in the dry-run script)

Once we are convinced that all changes are correct, we would then need
to update the original Dune queries so that we continue using the same
query ids.

---------

Co-authored-by: Felix Henneke <felix.henneke@protonmail.com>