Remove restriction on quote rewards for limit orders (#328)
When switching to all orders being limit orders we need to adapt how we
compute quote rewards. At the moment, only market orders are eligible
for quote rewards. With the current code this would mean that no quotes
would receive rewards.

This PR removes testing the kind of an order when querying for quote
solvers. This means that all quotes associated with executed orders are
elibible for rewards.

The impact of rewarding more quotes is expected to be small.

This change has to be done merged toether with a
[PR](https://github.com/cowprotocol/dune-sync/pull/59) on syncing quote
solvers to dune.

---------

Co-authored-by: harisang <harisangelidakis@gmail.com>
Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>