add price improvement fee
A new case was added to the batch rewards SQL query for computing a price improvement fee.
Price improvement is computed from a quote (depending on order kind). If
price improvement is positive and smaller than surplus, the protocol is
a fraction of price improvement. Otherwise it is zero or a fraction of
surplus, respectively.