Homogenize protocol fee queries (#354)
At the moment some of our infrastructure around solver payments is
duplicated across this repo and dune-sync. In particular, the code
blocks on protocol fees within sql queries are very similar between
queries.

This PR tries to adapt the protocol fee part of the batch reward query
in a way such that it can (essentially) be used as is in other queries.
This means that a small number of fields was added (e.g.
`protocol_fee_token` in the table `order_protocol_fee_prices`) which do
not serve a purpose in this repo but are neede in dune-sync.