Freeze numpy version (#365)
Numpy released a new major version 2.0.0. For some reason this makes our
code crash. This was noticed in #364 where unrelated tests started
failing.

I guess the problem is due to us using an outdated pandas version.

A quick fix is implemented in this PR. The version of numpy is fixed to
the latest release before 2.0.0.

This is not the first time we freeze a dependency to an old version.
Midterm we should update our dependencies instead.