Change use of block deadlines (#381)
This PR adds a grace period of one block to the deadline for settling on
chain. It also removes the use of block deadlines in computing fees and
other quantities.

The PR changes two things:
- Settlements are still rewarded if solvers settle one block after the
deadline. This change is implemented to not punish solvers for slightly
suboptimal transaction canceling in the driver. The alternative would be
to implement very cautios canceling which might negatively impact revert
rates.
- The use of block deadlines for setting `surplus`, `fee`, and
`execution_cost` to zero was removed. For every successful settlemen,
even if after the deadline by many blocks, those values will just be
copies of the respective values in the `settlement_observations` table.
The change to `surplus`and `execution_cost`does not have any effect on
solver payments. Those values are only used for analytics. The change to
`fee` has an effect since it enters the computation of
`network_fee_eth`. Before the change, the network fee was computed to be
negative and essentially equal to `-protocol_fee_eth`. On the one hand,
network fees were already removed from slippage using per order data. On
the other hand, slippage is also modified by adding the value of network
fees to reimburse solvers the cost of execution. Since network fees were
computed to be negative protocol fees, instead of paying solvers those
network fees they paid us protocol fees. See #380. With the proposed
change, network fees are always paid as part of slippage to the solver.
This is consistent with how slippage is handled in general.

The estimated impact on solver from adding a grace period is an increase
in overall rewards by about 1.48 ETH in the accounting period
2024-08-06--2024-08-13.
The impact of the second change on slippage is not estimated precisely.
The overcharging of protocol fees amounts to 0.6 ETH. The missing
reimbursement of network fees is not estimated.
See [this Dune query](https://dune.com/queries/3987946).

The changes should be mirrored to dune-sync and data for the accounting
period starting on 2024-08-13 should be re-synced.