Add service fee (#391)
This PR adds a service fee as per CIP-48.

The current implementation adds a boolean attribute `service_fee`to
`RewardAndPenaltyDatum`. This attribute is used in contructing payments
by scaling down rewards.

The query used on dune for determining the time in the bonding pool does
have a bug at the moment and does not account for the additional grace
period for solvers with a small bonding pool.

I tested the implementation for the previous and current accounting
period. The number in the previous period coincide. For the current
period, some solvers with a service fee did indeed receive a smaller
transfer.