Remove consistency rewards and dune parameters cleanup and fixing (#395)
As part of cleaning up and restructuring some Dune queries (see relevant
PRs [here](https://github.com/cowprotocol/dune-queries/pulls)), some
parameters needed to be dropped or renamed. This branch is what we used
to run the script for the payouts of Sep 3-10, 2024.

Moreover, we decided to remove consistency rewards altogether from the
main script, as we are moving away from those and it doesn't seem likely
that they will come back. Note however that the main sql query that
computes rewards, one of the core components of the script, is not
changed and consistency rewards are still computed there. This will be
addressed in a follow-up PR as that query is an important part of the
script and thus we thought it is better to address it separately.