Remove dead code (#413)
The PR removes some code which is not used at the moment.

- `token_list.py` is not a concept we currently use in our accounting.
We could add checks for that back into the accounting, but it looks more
like a circuit breaker job than the job of the accounting.
- `test_internal_trades.py` was testing internal trades, which is not
part of the accounting anymore. It has only been testing properties
which are not true anymore. Some of those tests could be refactored into
tests for slippage. But for now I think it is fine to remove it. I also
removed a reference in `test_models.py`.
- `dataset.py` was not used at all.
- `gap_detector.py` looks like interesting functionality. But it will
probably be superseded by whatever we build for dune-sync. We have not
used the functionality at least in the last year. So it should be fine
to remove it.
-

---------

Co-authored-by: Haris Angelidakis <64154020+harisang@users.noreply.github.com>